--- week_2/Assignment/AtHome/Error.go
@@ -1,9 +1,0 @@
-package main
-
-type customError struct {
-	message string
-}
-
-func (c customError) Error() string {
-	return c.message
-}
