+++ week_2/Assignment/AtHome/error.go
@@ -1,0 +1,21 @@
+package main
+
+type UnknownError struct {
+	message string
+}
+
+func (c UnknownError) Error() string {
+	return c.message
+}
+
+type ErrFileNotFound struct {
+
+}
+
+func (ErrFileNotFound) Error() string {
+
+}
+
+type ErrFileAccess struct{
+
+}
